requirements.txt
streamlit
openai
pymupdf
beautifulsoup4
requests
google-generativeai